Force import

+++ plugins/data_importer_exporter.rb
@@ -70,10 +70,12 @@ class DataImporterExporter < PluginBase
 
       # Skip import if checksum matches
       output_file = $OUTPUT_DIR + File.basename(files[i], ".yaml") + ".#{$DATA_TYPE}"
-      if File.exist?(output_file)
-        sha256 = Digest::SHA256.file output_file
-        firstLine = File.open($INPUT_DIR + files[i], &:readline)
-        next if firstLine[19..64+18] == sha256.hexdigest
+      unless $FORCE
+        if File.exist?(output_file)
+          sha256 = Digest::SHA256.file output_file
+          firstLine = File.open($INPUT_DIR + files[i], &:readline)
+          next if firstLine[19..64+18] == sha256.hexdigest
+        end
       end
 
       # Load the data from yaml file
@@ -145,7 +147,7 @@ class DataImporterExporter < PluginBase
     files = Dir.entries( $INPUT_DIR )
     files -= $DATA_IGNORE_LIST
     files = files.select { |e| File.extname(e) == ".#{$DATA_TYPE}" }
-    files = files.select { |e| file_modified_since?($INPUT_DIR + e, $STARTUP_TIME) or not data_file_exported?($INPUT_DIR + e) } unless $RE_EXPORT == true
+    files = files.select { |e| file_modified_since?($INPUT_DIR + e, $STARTUP_TIME) or not data_file_exported?($INPUT_DIR + e) } unless $FORCE == true
     files.sort!
  
     if files.empty?
@@ -176,14 +178,8 @@ class DataImporterExporter < PluginBase
  
       # Handle default values for the System data file
       if files[i] == "System.#{$DATA_TYPE}"
-        # Implement fix for RPG VXA - changed 'magic_number' to 'version_id'
-        if $DATA_TYPE == "rvdata2"
-            # Prevent the 'magic_number' field of System from always conflicting
-            data.version_id = $MAGIC_NUMBER unless $MAGIC_NUMBER == -1
-        else
-            # Prevent the 'magic_number' field of System from always conflicting
-            data.magic_number = $MAGIC_NUMBER unless $MAGIC_NUMBER == -1
-        end
+        # Prevent the 'magic_number' field of System from always conflicting
+        data.magic_number = $MAGIC_NUMBER unless $MAGIC_NUMBER == -1
         # Prevent the 'edit_map_id' field of System from conflicting
         data.edit_map_id = $DEFAULT_STARTUP_MAP unless $DEFAULT_STARTUP_MAP == -1
       end
